+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## v0.5.0
  - Change terminology in `steadybit service init` command to comply with terminology found in the public documentation.
  - Add a `steadybit service verify .steadybit.yml` command to compare the actual vs. desired resilience level.